# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeepResearch API",
    "description": "An agentic research assistant API for discovering and exploring AI/ML academic papers. Features a 3-agent pipeline (classifier \u2192 query optimiser \u2192 BGE vector search) powered by Google Gemini, with `BAAI/bge-base-en-v1.5` cosine similarity search over a local arXiv AI/ML corpus (~521k papers). Summarisation via Gemini. Built with FastAPI, SQLAlchemy, ChromaDB, and SQLite. Authenticate via the `X-API-Key` header on all endpoints except `/health`.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/community": {
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "List most popular papers",
        "description": "Returns papers ranked by community interaction count. Use the `period` parameter to filter by a rolling time window: `week` (last 7 days), `month` (last 30 days), `year` (last 365 days). Omit `period` for all-time rankings. A paper enters this index the first time it is directly accessed or summarised. Count increments on: direct paper lookup, summarisation, and related-papers queries.",
        "operationId": "list_community_papers_community_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of papers to return (max 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of papers to return (max 100)"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Period"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Rolling time window: week (7d), month (30d), year (365d). Omit for all-time.",
              "title": "Period"
            },
            "description": "Rolling time window: week (7d), month (30d), year (365d). Omit for all-time."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/community/{arxiv_id}": {
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "Get community stats for a paper",
        "description": "Returns the interaction count and last activity timestamp for a specific paper. A paper only appears here once it has been accessed via GET /papers/{arxiv_id}, POST /summarise (with arxiv_id), or GET /papers/{arxiv_id}/related.",
        "operationId": "get_community_paper_community__arxiv_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "arxiv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Arxiv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityPaperResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Paper has not been interacted with yet \u2014 not in the community index"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommunityListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of papers returned"
          },
          "papers": {
            "items": {
              "$ref": "#/components/schemas/CommunityPaperResponse"
            },
            "type": "array",
            "title": "Papers",
            "description": "Papers ranked by interaction count"
          }
        },
        "type": "object",
        "required": [
          "total",
          "papers"
        ],
        "title": "CommunityListResponse",
        "description": "Response from the community popular papers endpoint.",
        "examples": [
          {
            "papers": [
              {
                "arxiv_id": "1706.03762",
                "authors": "Ashish Vaswani, Noam Shazeer",
                "categories": "cs.CL cs.LG",
                "interaction_count": 14,
                "last_interacted_at": "2025-03-10T11:42:00Z",
                "title": "Attention Is All You Need",
                "url": "https://arxiv.org/abs/1706.03762",
                "year": 2017
              }
            ],
            "total": 1
          }
        ]
      },
      "CommunityPaperResponse": {
        "properties": {
          "arxiv_id": {
            "type": "string",
            "title": "Arxiv Id",
            "description": "arXiv paper ID"
          },
          "interaction_count": {
            "type": "integer",
            "title": "Interaction Count",
            "description": "Total number of times this paper has been accessed or summarised"
          },
          "last_interacted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Interacted At",
            "description": "Timestamp of the most recent interaction"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Paper title"
          },
          "authors": {
            "type": "string",
            "title": "Authors",
            "description": "Comma-separated author names"
          },
          "categories": {
            "type": "string",
            "title": "Categories",
            "description": "Space-separated arXiv categories"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year",
            "description": "Publication year"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "arXiv URL"
          }
        },
        "type": "object",
        "required": [
          "arxiv_id",
          "interaction_count",
          "last_interacted_at",
          "title",
          "authors",
          "categories",
          "url"
        ],
        "title": "CommunityPaperResponse",
        "description": "A paper entry in the community activity index.",
        "examples": [
          {
            "arxiv_id": "1706.03762",
            "authors": "Ashish Vaswani, Noam Shazeer",
            "categories": "cs.CL cs.LG",
            "interaction_count": 14,
            "last_interacted_at": "2025-03-10T11:42:00Z",
            "title": "Attention Is All You Need",
            "url": "https://arxiv.org/abs/1706.03762",
            "year": 2017
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Period": {
        "type": "string",
        "enum": [
          "week",
          "month",
          "year"
        ],
        "title": "Period"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}